Refactor configuration into a submodule
