+++ src/configuration/mod.rs
@@ -1,0 +1,8 @@
+mod configuration;
+mod common;
+mod client;
+mod dbus;
+
+pub use self::configuration::Configuration;
+pub use self::client::ClientConfiguration;
+pub use self::dbus::DBusConfiguration;
